# source_code/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: Village API
  version: 1.0.0
  description: Production-grade REST API for normalized Indian village-level geography data.
servers:
  - url: https://api.villageapi.com/v1
    description: Production
  - url: https://staging-api.villageapi.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local
security:
  - ApiKey: []
paths:
  /search:
    get:
      summary: Search villages
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: state
          schema:
            type: string
      responses:
        "200":
          description: Village dropdown options
        "400":
          description: Search query too short or invalid
        "401":
          description: API key missing or invalid
        "429":
          description: Rate limit exceeded
  /autocomplete:
    get:
      summary: Typeahead village suggestions
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Autocomplete suggestions
        "400":
          description: Search query too short or invalid
        "401":
          description: API key missing or invalid
  /states:
    get:
      summary: List all states
      responses:
        "200":
          description: States
  /states/{id}/districts:
    get:
      summary: List districts by state
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Districts for the selected state
  /districts/{id}/subdistricts:
    get:
      summary: List sub-districts by district
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sub-districts for the selected district
  /subdistricts/{id}/villages:
    get:
      summary: List villages by sub-district
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 500
      responses:
        "200":
          description: Villages for the selected sub-district
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    ApiSecret:
      type: apiKey
      in: header
      name: X-API-Secret